requirements.txt
numpy
Python-IO
pillow
keras
opencv-python-headless
streamlit
tensorflow
uvicorn
torch
torchvision
